+++ include/CppUtil/Engine/TriMesh.h
@@ -12,13 +12,13 @@ namespace CppUtil {
 		class TriMesh : public Primitive {
 			ELE_SETUP(TriMesh)
 		public:
-			TriMesh(const std::vector<size_t> & indice,
+			TriMesh(const std::vector<uint> & indice,
 				const std::vector<glm::vec3> & positions,
 				const std::vector<glm::vec3> & normals,
 				const std::vector<glm::vec3> & texcoords);
 
-			TriMesh(size_t triNum, size_t vertexNum,
-				const size_t * indice,
+			TriMesh(uint triNum, uint vertexNum,
+				const uint * indice,
 				const float * positions,
 				const float * normals,
 				const float * texcoords = nullptr);
@@ -31,11 +31,11 @@ namespace CppUtil {
 			const std::vector<glm::vec3> & GetPositions() const { return positions; }
 			const std::vector<glm::vec3> & GetNormals() const { return normals; }
 			const std::vector<glm::vec3> & GetTexcoords() const { return texcoords; }
-			const std::vector<size_t> & GetIndice() const { return indice; }
+			const std::vector<uint> & GetIndice() const { return indice; }
 			const std::vector<Triangle::Ptr> & GetTriangles() const { return triangles; }
 
 		private:
-			std::vector<size_t> indice;
+			std::vector<uint> indice;
 			std::vector<glm::vec3> positions;
 			std::vector<glm::vec3> normals;
 			std::vector<glm::vec3> texcoords;
