+++ src/CppUtil/Basic/Shape/Plane.cpp
@@ -14,15 +14,15 @@ Plane::Plane()
 
 	normalArr = new Array2D<float>(vertexNum, 3);
 	texCoordsArr = new Array2D<float>(vertexNum, 2);
-	indexArr = new Array2D<size_t>(triNum, 3);
+	indexArr = new Array2D<uint>(triNum, 3);
 	
-	for (size_t i = 0; i < vertexNum; i++) {
+	for (uint i = 0; i < vertexNum; i++) {
 		posArr->Copy(i, 0, 3, &(vertexData[8 * i]));
 		normalArr->Copy(i, 0, 3, &(vertexData[8 * i + 3]));
 		texCoordsArr->Copy(i, 0, 2, &(vertexData[8 * i + 6]));
 	}
 
-	const size_t indice[6] = {
+	const uint indice[6] = {
 		0, 1, 2,
 		3, 2, 1
 	};
@@ -54,21 +54,21 @@ float * Plane::GetTexCoordsArr() {
 	return texCoordsArr->GetData();
 }
 
-size_t * Plane::GetIndexArr() {
+uint * Plane::GetIndexArr() {
 	if (indexArr == nullptr)
 		return nullptr;
 
 	return indexArr->GetData();
 }
 
-size_t Plane::GetNormalArrSize() {
+uint Plane::GetNormalArrSize() {
 	return normalArr->GetMemSize();
 }
 
-size_t Plane::GetTexCoordsArrSize() {
+uint Plane::GetTexCoordsArrSize() {
 	return texCoordsArr->GetMemSize();
 }
 
-size_t Plane::GetIndexArrSize() {
+uint Plane::GetIndexArrSize() {
 	return indexArr->GetMemSize();
 }
