+++ src/CppUtil/Engine/Primitive/TriMesh.cpp
@@ -8,7 +8,7 @@ using namespace CppUtil::Basic;
 using namespace glm;
 using namespace std;
 
-TriMesh::TriMesh(const std::vector<size_t> & indice,
+TriMesh::TriMesh(const std::vector<uint> & indice,
 	const std::vector<glm::vec3> & positions,
 	const std::vector<glm::vec3> & normals,
 	const std::vector<glm::vec3> & texcoords)
@@ -24,12 +24,12 @@ TriMesh::TriMesh(const std::vector<size_t> & indice,
 
 	// traingel 的 mesh 在 init 的时候设置
 	// 因为现在还没有生成 share_ptr
-	for (size_t i = 0; i < indice.size(); i += 3)
+	for (uint i = 0; i < indice.size(); i += 3)
 		triangles.push_back(ToPtr(new Triangle(indice[i], indice[i+1], indice[i+2])));
 }
 
-TriMesh::TriMesh(size_t triNum, size_t vertexNum,
-	const size_t * indice,
+TriMesh::TriMesh(uint triNum, uint vertexNum,
+	const uint * indice,
 	const float * positions,
 	const float * normals,
 	const float * texcoords)
@@ -38,7 +38,7 @@ TriMesh::TriMesh(size_t triNum, size_t vertexNum,
 	assert(positions != nullptr);
 	assert(normals != nullptr);
 
-	for (size_t i = 0; i < vertexNum; i++) {
+	for (uint i = 0; i < vertexNum; i++) {
 		this->positions.push_back(vec3(positions[3 * i], positions[3 * i + 1], positions[3 * i + 2]));
 		this->normals.push_back(vec3(normals[3 * i], normals[3 * i + 1], normals[3 * i + 2]));
 		if (texcoords != nullptr)
@@ -47,7 +47,7 @@ TriMesh::TriMesh(size_t triNum, size_t vertexNum,
 
 	// traingel 的 mesh 在 init 的时候设置
 	// 因为现在还没有生成 share_ptr
-	for (size_t i = 0; i < triNum; i++) {
+	for (uint i = 0; i < triNum; i++) {
 		this->indice.push_back(indice[3 * i]);
 		this->indice.push_back(indice[3 * i + 1]);
 		this->indice.push_back(indice[3 * i + 2]);
